refactor: euclid, softmax; creates norm func.
